Minor tabulation and Doxygen comments fixes
 From http://www.stack.nl/~dimitri/doxygen/manual/docblocks.html in "Putting documentation after members":
"If you want to document the members of a file, struct, union, class, or enum, it is sometimes desired to place the documentation block after the member instead of before. For this purpose you have to put an additional < marker in the comment block. Note that this also works for the parameters of a function."